[BUGFIX] Webpage visibility setting change
Removes check "systemd unit present" in install.sh

The lines that used to check whether a systemd unit
is already in place ignored the following case:
    1. installed without --allow-remote-access
    2. try to reinstall to --allow-remote-access
NOT WORKING: the unit was not updated because
"already present".

TODO PROPOSAL: use rsync in place of cp, to
avoid copying the same file in the case reported
above.

ALTERNATIVE: save the address setting in the
config file, not in the systemd unit, if possible.